Updated use of onRowClicked in hoist-react and toolbox
