+++ client-app/src/mobile/grids/GridPage.js
@@ -1,4 +1,4 @@
-import {creates, hoistCmp, XH} from '@xh/hoist/core';
+import {creates, hoistCmp} from '@xh/hoist/core';
 import {panel} from '@xh/hoist/mobile/cmp/panel';
 import {grid} from '@xh/hoist/cmp/grid';
 import {filler} from '@xh/hoist/cmp/layout';
@@ -25,12 +25,7 @@ export const gridPage = hoistCmp.factory({
                     options: {prefix: 'Loaded'}
                 })
             ],
-            item: grid({
-                onRowClicked: (e) => {
-                    const {id} = e.data.raw;
-                    XH.appendRoute('gridDetail', {id});
-                }
-            }),
+            item: grid(),
             tbar: [
                 label('Size:'),
                 select({
