+++ client-app/src/mobile/treegrids/TreeGridPage.js
@@ -1,4 +1,4 @@
-import {creates, hoistCmp, XH} from '@xh/hoist/core';
+import {creates, hoistCmp} from '@xh/hoist/core';
 import {panel} from '@xh/hoist/mobile/cmp/panel';
 import {grid} from '@xh/hoist/cmp/grid';
 import {filler} from '@xh/hoist/cmp/layout';
@@ -15,12 +15,7 @@ export const treeGridPage = hoistCmp.factory({
             title: 'Tree Grids',
             icon: Icon.grid(),
             mask: 'onLoad',
-            item: grid({
-                onRowClicked: (e) => {
-                    const id = encodeURIComponent(e.data.raw.id);
-                    XH.appendRoute('treeGridDetail', {id});
-                }
-            }),
+            item: grid(),
             bbar: [
                 groupingChooser({maxWidth: 250}),
                 filler(),
